AdminDynamicBlog/requirements.txt

django
django-ckeditor
django-jazzmin
Pillow